+++ website/src/pages/elements/mod.rs
@@ -1,3 +1,5 @@
+mod button_page;
 mod tag_page;
 
+pub use button_page::ButtonPage;
 pub use tag_page::TagPage;