+++ website/src/pages/elements_page.rs
@@ -3,10 +3,7 @@ use leptos_bulma::elements::{BAButton, BBlock, BTitle};
 use leptos_bulma::layout::BSection;
 
 use crate::components::{GoToBulmaIo, GoToDocsRs, PageTitle};
-use crate::examples::{
-    BasicBlock, BasicBox, BasicButtons, BasicIcon, BasicTable, BasicTitle, ButtonColors,
-    ButtonSizes, ButtonStates, RustCodeExample,
-};
+use crate::examples::{BasicBlock, BasicBox, BasicIcon, BasicTable, BasicTitle, RustCodeExample};
 use crate::i18n::{t, use_i18n};
 
 #[component]
@@ -51,45 +48,9 @@ pub fn ElementsPage() -> impl IntoView {
                 "Button"
             </BTitle>
 
-            <BBlock>"Basic example:"</BBlock>
-
-            <RustCodeExample name="basic_buttons"/>
-
-            <BBlock>"See it in action:"</BBlock>
-
-            <BasicButtons/>
-
-            <BTitle is=5>"Colors"</BTitle>
-
-            <BBlock>"Example:"</BBlock>
-
-            <RustCodeExample name="button_colors"/>
-
-            <BBlock>"See it in action:"</BBlock>
-
-            <ButtonColors/>
-
-            <BTitle is=5>"Sizes"</BTitle>
-
-            <BBlock>"Example:"</BBlock>
-
-            <RustCodeExample name="button_sizes"/>
-
-            <BBlock>"See it in action:"</BBlock>
-
-            <ButtonSizes/>
-
-            <BTitle is=5>"States"</BTitle>
-
-            <BBlock>"Example:"</BBlock>
-
-            <RustCodeExample name="button_states"/>
-
-            <BBlock>"See it in action:"</BBlock>
-
-            <ButtonStates/>
-
-            <GoToDocsRs path="elements/fn.BButton"/>
+            <BAButton is_fullwidth=true href="/elements/button">
+                "Go to button page"
+            </BAButton>
         </BSection>
 
         <BSection>